Move StatePresentOffer
